FC-7_implementing_calculator_in_main_code_Evgen_change_ÐBigDecimal_to_double_001

+++ src/main/java/org/example/model/dto/FertilizerDTO.java
@@ -1,12 +1,10 @@
 package org.example.model.dto;
 
-import java.math.BigDecimal;
-
 public record FertilizerDTO (String fertilizerName,
                              String description,
-                             BigDecimal CaO,
-                             BigDecimal Mg,
-                             BigDecimal N,
-                             BigDecimal P,
-                             BigDecimal K) {
+                             double CaO,
+                             double Mg,
+                             double N,
+                             double P,
+                             double K) {
 }
