+++ src/main/java/org/example/model/dto/PeriodVegetationDTO.java
@@ -1,13 +1,11 @@
 package org.example.model.dto;
 
-import java.math.BigDecimal;
-
 public record PeriodVegetationDTO(String periodVegetationName,
                                   int numberOfDays,
                                   String description,
-                                  BigDecimal CaO,
-                                  BigDecimal Mg,
-                                  BigDecimal N,
-                                  BigDecimal P,
-                                  BigDecimal K) {
+                                  double CaO,
+                                  double Mg,
+                                  double N,
+                                  double P,
+                                  double K) {
 }
